+++ components.json
@@ -13,7 +13,7 @@
     "aliases": {
         "components": "~/components",
         "utils": "~/lib/utils",
-        "ui": "~/components/ui",
+        "ui": "~/components/backstage/ui",
         "lib": "~/lib",
         "hooks": "~/hooks"
     },
