+++ components/backstage/scenario/scenario-table.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import Link from 'next/link';
-import { ScenarioDeleteDialog } from '~/components/scenario/scenario-delete-dialog';
+import { ScenarioDeleteDialog } from '~/components/backstage/scenario/scenario-delete-dialog';
 import { type Scenario } from '~/lib/domain/scenario';
 import { ColumnDef } from '@tanstack/react-table';
-import { DataTable } from '~/components/ui/data-table';
+import { DataTable } from '~/components/backstage/ui/data-table';
 import { PropsWithoutRef } from 'react';
 import { Download, PlayIcon } from 'lucide-react';
