+++ components/backstage/theme/theme-toggle.tsx
@@ -4,13 +4,13 @@
 
 import { Moon, Sun } from 'lucide-react';
 import { useTheme } from 'next-themes';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/backstage/ui/button';
 import {
     DropdownMenu,
     DropdownMenuContent,
     DropdownMenuItem,
     DropdownMenuTrigger
-} from '@/components/ui/dropdown-menu';
+} from '@/components/backstage/ui/dropdown-menu';
 
 export function ThemeToggle() {
     const { setTheme } = useTheme();
