+++ components/backstage/ui/form.tsx
@@ -6,7 +6,7 @@ import { Slot } from '@radix-ui/react-slot';
 import { Controller, ControllerProps, FieldPath, FieldValues, FormProvider, useFormContext } from 'react-hook-form';
 
 import { cn } from '~/lib/utils';
-import { Label } from '~/components/ui/label';
+import { Label } from '~/components/backstage/ui/label';
 
 const Form = FormProvider;
