+++ components/backstage/ui/sidebar.tsx
@@ -7,12 +7,12 @@ import { PanelLeft } from 'lucide-react';
 
 import { useIsMobile } from '~/hooks/use-mobile';
 import { cn } from '~/lib/utils';
-import { Button } from '~/components/ui/button';
-import { Input } from '~/components/ui/input';
-import { Separator } from '~/components/ui/separator';
-import { Sheet, SheetContent } from '~/components/ui/sheet';
-import { Skeleton } from '~/components/ui/skeleton';
-import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '~/components/ui/tooltip';
+import { Button } from '~/components/backstage/ui/button';
+import { Input } from '~/components/backstage/ui/input';
+import { Separator } from '~/components/backstage/ui/separator';
+import { Sheet, SheetContent } from '~/components/backstage/ui/sheet';
+import { Skeleton } from '~/components/backstage/ui/skeleton';
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '~/components/backstage/ui/tooltip';
 
 const SIDEBAR_COOKIE_NAME = 'sidebar:state';
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7;
